Database.Orville.Conduit

Description

Documentation

selectConduit :: (Monad m, MonadOrville conn m, MonadCatch m, MonadResource m) => Select row -> ConduitT () row m ()

selectConduit provides a way to stream the results of a Select query from the database one by one using the conduit library. You can fuse the conduit built by this function with your own conduit pipeline to handle rows individually in whatever fashion you need (e.g. turning them into rows of CSV). This is useful if you want to be able to process many rows one by one without loading them all into memory at once. You can aggregate the results however you require as part of the conduit processing and then use runConduit (or runConduitRes) from the conduit library to execute the processing pipeline. Alternatively, your web server (wai, servant, etc) may provide support for converting a conduit into a streaming HTTP response.