Database.Orville.Popper

Description

Documentation

data PopError

Constructors

MissingRecord (TableDefinition readEntity writeEntity key) (FieldDefinition fieldValue) fieldValue

Unpoppable String

data Popper a b

data Popped a

Constructors

PoppedValue a

PoppedError PopError

(>>>) :: Category k cat => cat a b -> cat b c -> cat a c infixr 1

Left-to-right composition

(<<<) :: Category k cat => cat b c -> cat a b -> cat a c infixr 1

Right-to-left composition

abortPop :: PopError -> Popper a b

certainly :: PopError -> Popper (Maybe b) b

certainly' :: Popper a PopError -> Popper a (Maybe b) -> Popper a b

fromKern :: (a -> b) -> Popper a b

hasMany :: Ord fieldValue => TableDefinition readEntity writeEntity key -> FieldDefinition fieldValue -> Popper fieldValue [readEntity]

hasManyIn :: Ord fieldValue => TableDefinition readEntity writeEntity key -> FieldDefinition fieldValue -> Popper [fieldValue] (Map fieldValue [readEntity])

hasOneIn :: Ord fieldValue => TableDefinition readEntity writeEntity key -> FieldDefinition fieldValue -> Popper [fieldValue] (Map fieldValue readEntity)

hasManyInWhere :: Ord fieldValue => TableDefinition readEntity writeEntity key -> FieldDefinition fieldValue -> SelectOptions -> Popper [fieldValue] (Map fieldValue [readEntity])

hasManyWhere :: Ord fieldValue => TableDefinition readEntity writeEntity key -> FieldDefinition fieldValue -> SelectOptions -> Popper fieldValue [readEntity]

hasOne :: Ord fieldValue => TableDefinition readEntity writeEntity key -> FieldDefinition fieldValue -> Popper fieldValue (Maybe readEntity)

hasOne' :: Ord fieldValue => TableDefinition readEntity writeEntity key -> FieldDefinition fieldValue -> Popper fieldValue readEntity

hasOneWhere :: Ord fieldValue => TableDefinition readEntity writeEntity key -> FieldDefinition fieldValue -> SelectOptions -> Popper fieldValue (Maybe readEntity)

kern :: Popper a a

missingRecordMessage :: TableDefinition readEntity writeEntity key -> FieldDefinition fieldValue -> fieldValue -> String

popMissingRecord :: TableDefinition readEntity writeEntity key -> FieldDefinition fieldValue -> Popper fieldValue PopError

onKern :: (a -> b -> c) -> Popper b a -> Popper b c

pop :: MonadOrville conn m => Popper a b -> a -> m (Popped b)

popThrow :: MonadOrville conn m => Popper a b -> a -> m b

popFirst :: TableDefinition readEntity writeEntity key -> SelectOptions -> Popper a (Maybe readEntity)

popMany :: Popper a b -> Popper [a] [b]

onPopMany :: Popper a b -> Popper [a] [b] -> Popper a b

popMaybe :: Popper a b -> Popper (Maybe a) (Maybe b)

popQuery :: String -> (forall conn m. MonadOrville conn m => m b) -> Popper a b

popQuery embeds an Orville operation in a popper. It is left up to the programmer to ensure that the Orville operation does not do any updates to the database, but only does queries.

The initial string argument is a description of the query to put into the results of explain

popRecord :: TableDefinition readEntity writeEntity key -> key -> Popper a (Maybe readEntity)

popRecord' :: TableDefinition readEntity writeEntity key -> key -> Popper a readEntity

popTable :: TableDefinition readEntity writeEntity key -> SelectOptions -> Popper a [readEntity]

explain :: Popper a b -> String

explainLines :: Popper a b -> [String]