Database.Orville.Core

Description

Documentation

data TableDefinition readEntity writeEntity key

A TableDefinition is the center of the Orville universe. A TableDefinition defines the structure of a table in the database and associates it with a Haskell datatype, usually a Haskell record type. The TableDefinition must specify how the Haskell type is converted to and from the database schema, as as well as provide same basic utility functions required by Orville for interacting with the Haskell datatype.

Usually you will use TableParams to construct a TableDefinition in a more concise way. This type is provided as an escape hatch for any situations where TableParams is too restrictive for the sql mapping required by a type.

Constructors

TableDefinition

Fields

tableName :: String

The name of the table in the database.

tableFields :: [SomeField]

A list of field definitions defining the table structure

tableSafeToDelete :: [String]

A list of any columns that may be deleted from the table by Orville. (Orville will never delete a column without being told it is safe)

tablePrimaryKey :: FieldDefinition key

The statically typed field definition that is the primary key. Currently this field must still by listed in tableFields

tableFromSql :: FromSql readEntity

A definition of how to convert the haskell type from a sql row

tableToSql :: ToSql writeEntity ()

A function to set the key on the entity

tableGetKey :: readEntity -> key

A function to get the key on the entity

tableComments :: TableComments ()

Any comments that might be interesting for developers to see. These comments will get printed in the log if there is an erro while attempting to migrate the table.

mkTableDefinition :: TableParams readEntity writeEntity key -> TableDefinition readEntity writeEntity key

mkTableDefinition converts a TableParams to TableDefinition. Usually this is used directly on a record literal of the TableParams. For example:

 data Foo key = Foo key { fooId :: Record }
 myTable :: TableDefinition Foo
 myTable = mkTableDefinition $
   TableParams
     { tblName = "foo"
     , tblMapper = User $ attrField fooId idField
     , tableSafeToDelete = []
     , tblSetKey = key foo -> foo { fooId = key }
     , tblGetKey = fooId
     , tblComments = []
     }

tableKeyToSql :: TableDefinition readEntity writeEntity key -> key -> SqlValue

tableKeyFromSql :: TableDefinition readEntity writeEntity key -> SqlValue -> Maybe key

data SqlType a

SqlType defines the mapping of a Haskell type (a) to a SQL column type in the database. This includes both how to convert the type to and from the raw values read from the database as well as the schema information required to create and migrate columns using the type.

Constructors

SqlType

Fields

sqlTypeDDL :: String

The raw SQL DDL to use when creating/migrating columns of this type (not including any NULL or NOT NULL declarations)

sqlTypeReferenceDDL :: Maybe String

The raw SQL DDL to use when creating/migrating columns with foreign keys to this type. This is used foreignRefType to build a new SqlType when making foreign key fields

sqlTypeNullable :: Bool

Indicates whether columns should be marked NULL or NOT NULL in the database schema. If this is True, then sqlTypeFromSql should provide a handling of SqlNull that returns an a, not Nothing.

sqlTypeId :: SqlTypeId

sqlTypeId will be compared to the colType field found in the SqlColDesc return by describeTable when determining whether a column type change is required when migrating the database.

sqlTypeSqlSize :: Maybe Int

'sqlTypeSqlSize will be compared to the colSize field found in the SqlColDesc return by describeTable when determining whether a column type change is required when migrating the database.

sqlTypeToSql :: a -> SqlValue

A function for converting Haskell values of this type into values to be stored in the database.

sqlTypeFromSql :: SqlValue -> Maybe a

A function for converting values of this are stored in the database into Haskell values. This function should return Nothing to indicate an error if the conversion is impossible. Otherwise it should return Just the corresponding a value.

serial :: SqlType Int32

serial defines a 32-bit auto-incrementing column type. This corresponds to the SERIAL type in PostgresSQL.

bigserial :: SqlType Int64

serial defines a 64-bit auto-incrementing column type. This corresponds to the BIGSERIAL type in PostgresSQL.

text :: Int -> SqlType Text

text defines a fixed length text field type. This corresponds to a "CHAR(len)" type in SQL.

varText :: Int -> SqlType Text

varText defines a variable text field type with a max length. This corresponds to a "VARCHAR(len)" type in SQL.

integer :: SqlType Int32

integer defines a 32-bit integer type. This corresponds to the INTEGER type in SQL.

bigInteger :: SqlType Int64

bigInteger defines a 64-bit integer type. This corresponds to the BIGINT type in SQL.

double :: SqlType Double

double defines a floating point numeric type. This corresponds to the "DOUBLE PRECISION" type in SQL.

boolean :: SqlType Bool

boolean defines a True/False boolean type. This corresponds to the BOOLEAN type in SQL.

date :: SqlType Day

date defines a type representing a calendar date (without time zone). It corresponds to the DATE type in SQL.

timestamp :: SqlType UTCTime

timestamp defines a type representing a particular point in time (without time zone). It corresponds to the "TIMESTAMP with time zone" type in SQL.

Note: This is NOT a typo. The "TIMESTAMP with time zone" type in SQL does not include any actual time zone information. For an excellent explanation of the complexities involving this type, please see Chris Clark's blog post about it: http://blog.untrod.com/2016/08/actually-understanding-timezones-in-postgresql.html

textSearchVector :: SqlType Text

textSearchVector defines a type for indexed text searching. It corresponds to the TSVECTOR type in PostgresSQL.

convertSqlType :: (b -> a) -> (a -> b) -> SqlType a -> SqlType b

convertSqlType changes the Haskell type used by a SqlType in the same manner as maybeConvertSqlType in cases where an a can always be converted to a b.

maybeConvertSqlType :: (b -> a) -> (a -> Maybe b) -> SqlType a -> SqlType b

maybeConvertSqlType changes the Haskell type used by a SqlType which changing the column type that will be used in the database schema. The functions given will be used to convert the now Haskell type to and from the original type when reading and writing values from the database. When reading an a value from the database, the conversion function should produce Nothing if the value cannot be successfully converted to a b

nullableType :: SqlType a -> SqlType (Maybe a)

nullableType creates a nullable version of an existing SqlType. The underlying sql type will be the same as the original, but column will be created with a NULL constraint instead a 'NOT NULL' constraint. The Haskell value Nothing will be used represent NULL values when converting to and from sql.

data TableParams readEntity writeEntity key

TableParams is the simplest way to make a TableDefinition. You can use mkTableDefinition to make a definition from the simplified params. Where TableDefinition requires the tableFields, tableFromSql, and tableToSql to all be defined separately and kept in sync, TableParams provides a single tblMapper field that specifies all three simultaneously and ensures they are consistent with one another.

Constructors

TableParams

Fields

tblName :: String

The name of the table in the database

tblMapper :: RelationalMap writeEntity readEntity

The relational mapping that defines how the Haskell entity type is converted both to and from sql. The fields utilized in the mapping are used to automatically build the list of FieldDefinitions that define the structure of the table in the database.

tblSafeToDelete :: [String]

A list of any columns that may be deleted from the table by Orville. (Orville will never delete a column without being told it is safe)

tblPrimaryKey :: FieldDefinition key

A function to set the key on the entity

tblGetKey :: readEntity -> key

A function to get the key on the entity

tblComments :: TableComments ()

Any comments that might be interesting for developers to see. These comments will get printed in the log if there is an erro while attempting to migrate the table.

data RelationalMap a b

mapAttr :: (a -> b) -> RelationalMap b c -> RelationalMap a c

mapField :: FieldDefinition a -> RelationalMap a a

attrField :: (a -> b) -> FieldDefinition b -> RelationalMap a b

maybeMapper :: RelationalMap a b -> RelationalMap (Maybe a) (Maybe b)

prefixMap :: String -> RelationalMap a b -> RelationalMap a b

partialMap :: RelationalMap a (Either String a) -> RelationalMap a a

readOnlyMap :: RelationalMap a b -> RelationalMap c b

readOnlyField :: FieldDefinition a -> RelationalMap b a

data OrvilleEnv conn

OrvilleEnv tracks all the environment information required for an 'OrvilleT conn m' Monad to operate. Use newOrvilleEnv to construct one.

newOrvilleEnv :: Pool conn -> OrvilleEnv conn

newOrvilleEnv initialized an OrvilleEnv for service. The connection pool provided will be used to obtain connections to the database ase required. You can use the createConnectionPool utility function to create a connection pool to a PosgreSQL server.

setStartTransactionSQL :: String -> OrvilleEnv conn -> OrvilleEnv conn

aroundRunningQuery :: (forall a. QueryType -> String -> IO a -> IO a) -> OrvilleEnv conn -> OrvilleEnv conn

addTransactionCallBack :: (TransactionEvent -> IO ()) -> OrvilleEnv conn -> OrvilleEnv conn

ormEnvPool :: OrvilleEnv conn -> Pool conn

data TransactionEvent

Constructors

TransactionStart

TransactionCommit

TransactionRollback

data OrvilleT conn m a

unOrvilleT :: OrvilleT conn m a -> ReaderT (OrvilleEnv conn) m a

data SqlValue :: *

class (Monad m, MonadIO m, IConnection conn, MonadBaseControl IO m, MonadThrow m) => MonadOrville conn m | m -> conn where

Minimal complete definition

getOrvilleEnv, localOrvilleEnv

Methods

getOrvilleEnv :: m (OrvilleEnv conn)

localOrvilleEnv :: (OrvilleEnv conn -> OrvilleEnv conn) -> m a -> m a

runOrville :: OrvilleT conn m a -> OrvilleEnv conn -> m a

mapOrvilleT :: Monad n => (m a -> n b) -> OrvilleT conn m a -> OrvilleT conn n b

data QueryType

Constructors

SelectQuery

InsertQuery

UpdateQuery

DeleteQuery

DDLQuery

withTransaction :: MonadOrville conn m => m a -> m a

data ColumnFlag

Constructors

PrimaryKey

ColumnDefault a => Default a

Unique

References (TableDefinition readEntity writeEntity key) (FieldDefinition key)

ColumnDescription String

AssignedByDatabase

data Now

Constructors

Now

data ColumnType

Constructors

AutomaticId

ForeignId

Text Int

VarText Int

Date

Timestamp

Integer

BigInteger

TextSearchVector

Double

Boolean

data FieldDefinition a

fieldOfType :: SqlType a -> String -> FieldDefinition a

textField :: String -> Int -> FieldDefinition Text

fixedTextField :: String -> Int -> FieldDefinition Text

dayField :: String -> FieldDefinition Day

utcTimeField :: String -> FieldDefinition UTCTime

int32Field :: String -> FieldDefinition Int32

int64Field :: String -> FieldDefinition Int64

doubleField :: String -> FieldDefinition Double

boolField :: String -> FieldDefinition Bool

automaticIdField :: String -> FieldDefinition Int32

searchVectorField :: String -> FieldDefinition Text

nullableField :: FieldDefinition a -> FieldDefinition (Maybe a)

foreignKeyField :: String -> TableDefinition readEntity writeEntity key -> FieldDefinition key -> FieldDefinition key

withFlag :: FieldDefinition a -> ColumnFlag -> FieldDefinition a

withName :: FieldDefinition a -> String -> FieldDefinition a

withConversion :: FieldDefinition a -> (SqlType a -> SqlType b) -> FieldDefinition b

fieldFromSql :: FieldDefinition a -> FromSql a

fieldToSqlValue :: FieldDefinition a -> a -> SqlValue

data SomeField

Constructors

SomeField (FieldDefinition a)

withPrefix :: FieldDefinition a -> String -> FieldDefinition a

fieldName :: FieldDefinition a -> String

fieldType :: FieldDefinition a -> SqlType a

fieldFlags :: FieldDefinition a -> [ColumnFlag]

data IndexDefinition

Constructors

IndexDefinition

Fields

indexName :: String

indexUnique :: Bool

indexTable :: String

indexBody :: String

uniqueIndex :: String -> TableDefinition readEntity writeEntity key -> [SomeField] -> IndexDefinition

simpleIndex :: String -> TableDefinition readEntity writeEntity key -> [SomeField] -> IndexDefinition

data ConstraintDefinition

Constructors

ConstraintDefinition

Fields

constraintName :: String

constraintTable :: String

constraintBody :: String

uniqueConstraint :: String -> TableDefinition readEntity writeEntity key -> [SomeField] -> ConstraintDefinition

dropConstraint :: TableDefinition readEntity writeEntity key -> String -> SchemaItem

data FromSql a

data FromSqlError

Constructors

RowDataError String

QueryError String

class ColumnSpecifier col where

Minimal complete definition

selectForm

Methods

selectForm :: col -> SelectForm

col :: (ColumnSpecifier col, Convertible SqlValue a) => col -> FromSql a

data ToSql a b

getField :: Convertible a SqlValue => (entity -> a) -> ToSql entity ()

getComponent :: (entity -> a) -> ToSql a () -> ToSql entity ()

data SchemaItem

Constructors

Table (TableDefinition readEntity writeEntity key)

DropTable String

Index IndexDefinition

DropIndex String

Constraint ConstraintDefinition

DropConstraint String String

type SchemaDefinition = [SchemaItem]

type Record = Int

type CreatedAt = UTCTime

type UpdatedAt = UTCTime

type OccurredAt = UTCTime

data TableComments a

noComments :: TableComments ()

say :: String -> (Int, Int, Int) -> String -> TableComments ()

data WhereCondition

whereAnd :: [WhereCondition] -> WhereCondition

whereOr :: [WhereCondition] -> WhereCondition

whereIn :: FieldDefinition a -> [a] -> WhereCondition

whereLike :: FieldDefinition a -> String -> WhereCondition

whereLikeInsensitive :: FieldDefinition a -> String -> WhereCondition

whereNotIn :: FieldDefinition a -> [a] -> WhereCondition

whereQualified :: TableDefinition a b c -> WhereCondition -> WhereCondition

whereRaw :: String -> [SqlValue] -> WhereCondition

isNull :: FieldDefinition a -> WhereCondition

isNotNull :: FieldDefinition a -> WhereCondition

(.==) :: FieldDefinition a -> a -> WhereCondition

(.<>) :: FieldDefinition a -> a -> WhereCondition

(.<-) :: FieldDefinition a -> [a] -> WhereCondition

(%==) :: FieldDefinition a -> a -> WhereCondition

(.>) :: FieldDefinition a -> a -> WhereCondition

(.>=) :: FieldDefinition a -> a -> WhereCondition

(.<) :: FieldDefinition a -> a -> WhereCondition

(.<=) :: FieldDefinition a -> a -> WhereCondition

data SelectOptions

Constructors

SelectOptions

Fields

selectDistinct :: First Bool

selectOptWhere :: [WhereCondition]

selectOptOrder :: [OrderByClause]

selectOptLimit :: First Int

selectOptOffset :: First Int

selectOptGroup :: [GroupByClause]

where_ :: WhereCondition -> SelectOptions

distinct :: SelectOptions

order :: ToOrderBy a => a -> SortDirection -> SelectOptions

limit :: Int -> SelectOptions

offset :: Int -> SelectOptions

groupBy :: ToGroupBy a => a -> SelectOptions

(<>) :: Monoid m => m -> m -> m infixr 6

An infix synonym for mappend.

Since: 4.5.0.0

data FieldUpdate

fieldUpdate :: FieldDefinition a -> a -> FieldUpdate

(.:=) :: FieldDefinition a -> a -> FieldUpdate

data OrderByClause

Constructors

OrderByClause String [SqlValue] SortDirection

data SortDirection

Constructors

Ascending

Descending

migrateSchema :: MonadOrville conn m => SchemaDefinition -> m ()

migrateSchema will attempt to make changes to the actual database schema that it it matches the provided SchemaDefinition. Unsafe migrations such as dropping tables or columns are never attempted unless the SchemaDefinition explicitly states that the items are safe to drop. Column types may be changed, but will fail if the database cannot successfully make the request type change.

data MigrationError

Constructors

MigrationLockExcessiveRetryError String

MigrationExecutionError SchemaItem SomeException

generateMigrationPlan :: MonadOrville conn m => SchemaDefinition -> m (Maybe MigrationPlan)

generateMigrationPlan inspects the state of the actual database schema and constructs a plan describing what changes would be made to make it match the provided SchemaDefinition. If the actual schema already matches the definition, Nothing will be returned.

data MigrationPlan

data MigrationItem

Constructors

MigrationItem

Fields

migrationItemSchemaItem :: SchemaItem

migrationItemDDL :: DDL

migrationPlanItems :: MigrationPlan -> [MigrationItem]

selectAll :: MonadOrville conn m => TableDefinition readEntity writeEntity key -> SelectOptions -> m [readEntity]

selectFirst :: MonadOrville conn m => TableDefinition readEntity writeEntity key -> SelectOptions -> m (Maybe readEntity)

deleteRecord :: MonadOrville conn m => TableDefinition readEntity writeEntity key -> key -> m ()

deleteWhere :: MonadOrville conn m => TableDefinition readEntity writeEntity key -> [WhereCondition] -> m Integer

findRecord :: MonadOrville conn m => TableDefinition readEntity writeEntity key -> key -> m (Maybe readEntity)

findRecords :: (Ord key, MonadOrville conn m) => TableDefinition readEntity writeEntity key -> [key] -> m (Map key readEntity)

findRecordsBy :: (Ord fieldValue, MonadOrville conn m) => TableDefinition readEntity writeEntity key -> FieldDefinition fieldValue -> SelectOptions -> m (Map fieldValue [readEntity])

insertRecord :: MonadOrville conn m => TableDefinition readEntity writeEntity key -> writeEntity -> m readEntity

insertRecordMany :: MonadOrville conn m => TableDefinition readEntity writeEntity key -> [writeEntity] -> m ()

updateFields :: MonadOrville conn m => TableDefinition readEntity writeEntity key -> [FieldUpdate] -> [WhereCondition] -> m Integer

updateRecord :: MonadOrville conn m => TableDefinition readEntity writeEntity key -> key -> writeEntity -> m ()