Database.Orville.Core

Description

Documentation

data TableDefinition entity

A TableDefinition is the center of the Orville universe. A TableDefinition defines the structure of a table in the database and associates it with a Haskell datatype, usually a Haskell record type. The TableDefinition must specify how the Haskell type is converted to and from the database schema, as as well as provide same basic utility functions required by Orville for interacting with the Haskell datatype.

Usually you will use TableParams to construct a TableDefinition in a more concise way. This type is provided as an escape hatch for any situations where TableParams is too restrictive for the sql mapping required by a type.

Constructors

TableDefinition

Fields

tableName :: String

The name of the table in the database.

tableFields :: [FieldDefinition]

A list of field definitions defining the table structure

tableSafeToDelete :: [String]

A list of any columns that may be deleted from the table by Orville. (Orville will never delete a column without being told it is safe)

tableFromSql :: FromSql (entity Record)

A definition of how to convert the haskell type from a sql row

tableToSql :: forall key. ToSql (entity key) ()

A definition of how to convert the haskell type to a sql row

tableSetKey :: forall key1 key2. key2 -> entity key1 -> entity key2

A function to set the key on the entity

tableGetKey :: forall key. entity key -> key

A function to get the key on the entity

tableComments :: TableComments ()

Any comments that might be interesting for developers to see. These comments will get printed in the log if there is an erro while attempting to migrate the table.

mkTableDefinition :: TableParams entity -> TableDefinition entity

mkTableDefinition converts a TableParams to TableDefinition. Usually this is used directly on a record literal of the TableParams. For example:

 data Foo key = Foo key { fooId :: Record }
 myTable :: TableDefinition Foo
 myTable = mkTableDefinition $
   TableParams
     { tblName = "foo"
     , tblMapper = User $ attrField fooId idField
     , tableSafeToDelete = []
     , tblSetKey = key foo -> foo { fooId = key }
     , tblGetKey = fooId
     , tblComments = []
     }

data TableParams entity

TableParams is the simplest way to make a TableDefinition. You can use mkTableDefinition to make a definition from the simplified params. Where TableDefinition requires the tableFields, tableFromSql, and tableToSql to all be defined separately and kept in sync, TableParams provides a single tblMapper field that specifies all three simultaneously and ensures they are consistent with one another.

Constructors

TableParams

Fields

tblName :: String

The name of the table in the database

tblMapper :: RelationalMap (entity Record) (entity Record)

The relational mapping that defines how the Haskell entity type is converted both to and from sql. The fields utilized in the mapping are used to automatically build the list of FieldDefinitions that define the structure of the table in the database.

tblSafeToDelete :: [String]

A list of any columns that may be deleted from the table by Orville. (Orville will never delete a column without being told it is safe)

tblSetKey :: forall key1 key2. key2 -> entity key1 -> entity key2

A function to set the key on the entity

tblGetKey :: forall key. entity key -> key

A function to get the key on the entity

tblComments :: TableComments ()

Any comments that might be interesting for developers to see. These comments will get printed in the log if there is an erro while attempting to migrate the table.

data RelationalMap a b

mapAttr :: (a -> b) -> RelationalMap b c -> RelationalMap a c

mapField :: (Convertible a SqlValue, Convertible SqlValue a) => FieldDefinition -> RelationalMap a a

attrField :: (Convertible b SqlValue, Convertible SqlValue b) => (a -> b) -> FieldDefinition -> RelationalMap a b

maybeMapper :: RelationalMap a b -> RelationalMap (Maybe a) (Maybe b)

prefixMap :: String -> RelationalMap a b -> RelationalMap a b

partialMap :: RelationalMap a (Either String a) -> RelationalMap a a

readOnlyMap :: RelationalMap a b -> RelationalMap c b

tablePrimaryKey :: TableDefinition entity -> FieldDefinition

data OrvilleEnv conn

OrvilleEnv tracks all the environment information required for an 'OrvilleT conn m' Monad to operate. Use newOrvilleEnv to construct one.

newOrvilleEnv :: Pool conn -> OrvilleEnv conn

newOrvilleEnv initialized an OrvilleEnv for service. The connection pool provided will be used to obtain connections to the database ase required. You can use the createConnectionPool utility function to create a connection pool to a PosgreSQL server.

ormEnvPool :: OrvilleEnv conn -> Pool conn

type Orville a = forall m conn. (MonadFail m, MonadOrville conn m, MonadThrow m) => m a

data OrvilleT conn m a

unOrvilleT :: OrvilleT conn m a -> ReaderT (OrvilleEnv conn) m a

data SqlValue

SqlValue is the main type for expressing Haskell values to SQL databases.

INTRODUCTION TO SQLVALUE

This type is used to marshall Haskell data to and from database APIs. HDBC driver interfaces will do their best to use the most accurate and efficient way to send a particular value to the database server.

Values read back from the server are constructed with the most appropriate SqlValue constructor. fromSql or safeFromSql can then be used to convert them into whatever type is needed locally in Haskell.

Most people will use toSql and fromSql instead of manipulating SqlValues directly.

EASY CONVERSIONS BETWEEN HASKELL TYPES

Conversions are powerful; for instance, you can call fromSql on a SqlInt32 and get a String or a Double out of it. This class attempts to Do The Right Thing whenever possible, and will raise an error when asked to do something incorrect. In particular, when converting to any type except a Maybe, SqlNull as the input will cause an error to be raised.

Conversions are implemented in terms of the Data.Convertible module, part of the convertible package. You can refer to its documentation, and import that module, if you wish to parse the Left result from safeFromSql yourself, or write your own conversion instances.

Here are some notes about conversion:

Fractions of a second are not preserved on time values

There is no safeToSql because toSql never fails.

See also toSql, safeFromSql, fromSql, nToSql, iToSql, posixToSql.

ERROR CONDITIONS

There may sometimes be an error during conversion. For instance, if you have a SqlString and are attempting to convert it to an Integer, but it doesn't parse as an Integer, you will get an error. This will be indicated as an exception if using fromSql, or a Left result if using safeFromSql.

SPECIAL NOTE ON POSIXTIME

Note that a NominalDiffTime or POSIXTime is converted to SqlDiffTime by toSql. HDBC cannot differentiate between NominalDiffTime and POSIXTime since they are the same underlying type. You must construct SqlPOSIXTime manually or via posixToSql, or use SqlUTCTime.

DETAILS ON SQL TYPES

HDBC database backends are expected to marshal date and time data back and forth using the appropriate representation for the underlying database engine. Databases such as PostgreSQL with builtin date and time types should see automatic conversion between these Haskell types to database types. Other databases will be presented with an integer or a string. Care should be taken to use the same type on the Haskell side as you use on the database side. For instance, if your database type lacks timezone information, you ought not to use ZonedTime, but instead LocalTime or UTCTime. Database type systems are not always as rich as Haskell. For instance, for data stored in a TIMESTAMP WITHOUT TIME ZONE column, HDBC may not be able to tell if it is intended as UTCTime or LocalTime data, and will happily convert it to both, upon your request. It is your responsibility to ensure that you treat timezone issues with due care.

This behavior also exists for other types. For instance, many databases do not have a Rational type, so they will just use the show function and store a Rational as a string.

The conversion between Haskell types and database types is complex, and generic code in HDBC or its backends cannot possibly accomodate every possible situation. In some cases, you may be best served by converting your Haskell type to a String, and passing that to the database.

UNICODE AND BYTESTRINGS

Beginning with HDBC v2.0, interactions with a database are presumed to occur in UTF-8.

To accomplish this, whenever a ByteString must be converted to or from a String, the ByteString is assumed to be in UTF-8 encoding, and will be decoded or encoded as appropriate. Database drivers will generally present text or string data they have received from the database as a SqlValue holding a ByteString, which fromSql will automatically convert to a String, and thus automatically decode UTF-8, when you need it. In the other direction, database drivers will generally convert a SqlString to a ByteString in UTF-8 encoding before passing it to the database engine.

If you are handling some sort of binary data that is not in UTF-8, you can of course work with the ByteString directly, which will bypass any conversion.

Due to lack of support by database engines, lazy ByteStrings are not passed to database drivers. When you use toSql on a lazy ByteString, it will be converted to a strict ByteString for storage. Similarly, fromSql will convert a strict ByteString to a lazy ByteString if you demand it.

EQUALITY OF SQLVALUE

Two SqlValues are considered to be equal if one of these hold. The first comparison that can be made is controlling; if none of these comparisons can be made, then they are not equal:

Both are NULL

Both represent the same type and the encapsulated values are considered equal by applying (==) to them

The values of each, when converted to a string, are equal

STRING VERSIONS OF TIMES

Default string representations are given as comments below where such are non-obvious. These are used for fromSql when a String is desired. They are also defaults for representing data to SQL backends, though individual backends may override them when a different format is demanded by the underlying database. Date and time formats use ISO8601 date format, with HH:MM:SS added for time, and -HHMM added for timezone offsets.

DEPRECATED CONSTRUCTORS

SqlEpochTime and SqlTimeDiff are no longer created automatically by any toSql or fromSql functions or database backends. They may still be manually constructed, but are expected to be removed in a future version. Although these two constructures will be removed, support for marshalling to and from the old System.Time data will be maintained as long as System.Time is, simply using the newer data types for conversion.

class (Monad m, MonadIO m, IConnection conn, MonadBaseControl IO m) => MonadOrville conn m | m -> conn where

Minimal complete definition

getOrvilleEnv, localOrvilleEnv

Methods

getOrvilleEnv :: m (OrvilleEnv conn)

localOrvilleEnv :: (OrvilleEnv conn -> OrvilleEnv conn) -> m a -> m a

startTransactionSQL :: m String

runningQuery :: QueryType -> String -> m a -> m a

runOrville :: OrvilleT conn m a -> OrvilleEnv conn -> m a

mapOrvilleT :: Monad n => (m a -> n b) -> OrvilleT conn m a -> OrvilleT conn n b

data QueryType

Constructors

SelectQuery

InsertQuery

UpdateQuery

DeleteQuery

DDLQuery

withTransaction :: MonadOrville conn m => m a -> m a

data ColumnFlag

Constructors

PrimaryKey

ColumnDefault a => Default a

Null

Unique

References (TableDefinition entity) FieldDefinition

ColumnDescription String

data Now

Constructors

Now

data ColumnType

Constructors

AutomaticId

ForeignId

Text Int

VarText Int

Date

Timestamp

Integer

BigInteger

TextSearchVector

Double

Boolean

type FieldDefinition = (String, ColumnType, [ColumnFlag])

withPrefix :: FieldDefinition -> String -> FieldDefinition

fieldName :: FieldDefinition -> String

data IndexDefinition

Constructors

IndexDefinition

Fields

indexName :: String

indexUnique :: Bool

indexTable :: String

indexBody :: String

uniqueIndex :: String -> TableDefinition entity -> [FieldDefinition] -> IndexDefinition

simpleIndex :: String -> TableDefinition entity -> [FieldDefinition] -> IndexDefinition

data ConstraintDefinition

Constructors

ConstraintDefinition

Fields

constraintName :: String

constraintTable :: String

constraintBody :: String

uniqueConstraint :: String -> TableDefinition entity -> [FieldDefinition] -> ConstraintDefinition

dropConstraint :: TableDefinition entity -> String -> SchemaItem

data FromSql a

data FromSqlError

Constructors

RowDataError String

QueryError String

class ColumnSpecifier col where

Methods

selectForm :: col -> SelectForm

col :: (ColumnSpecifier col, Convertible SqlValue a) => col -> FromSql a

data ToSql a b

getField :: Convertible a SqlValue => (entity -> a) -> ToSql entity ()

getComponent :: (entity -> a) -> ToSql a () -> ToSql entity ()

withFlag :: FieldDefinition -> ColumnFlag -> FieldDefinition

withName :: FieldDefinition -> String -> FieldDefinition

data SchemaItem

Constructors

Table (TableDefinition entity)

DropTable String

Index IndexDefinition

DropIndex String

Constraint ConstraintDefinition

DropConstraint String String

type SchemaDefinition = [SchemaItem]

type Record = Int

type CreatedAt = UTCTime

type UpdatedAt = UTCTime

type OccurredAt = UTCTime

data TableComments a

noComments :: TableComments ()

say :: String -> (Int, Int, Int) -> String -> TableComments ()

data WhereCondition

whereAnd :: [WhereCondition] -> WhereCondition

whereOr :: [WhereCondition] -> WhereCondition

whereIn :: FieldDefinition -> [SqlValue] -> WhereCondition

whereNotIn :: FieldDefinition -> [SqlValue] -> WhereCondition

isNull :: FieldDefinition -> WhereCondition

isNotNull :: FieldDefinition -> WhereCondition

(.==) :: Convertible a SqlValue => FieldDefinition -> a -> WhereCondition

(.<>) :: Convertible a SqlValue => FieldDefinition -> a -> WhereCondition

(.<-) :: Convertible a SqlValue => FieldDefinition -> [a] -> WhereCondition

(%==) :: Convertible a SqlValue => FieldDefinition -> a -> WhereCondition

(.>) :: Convertible a SqlValue => FieldDefinition -> a -> WhereCondition

(.>=) :: Convertible a SqlValue => FieldDefinition -> a -> WhereCondition

(.<) :: Convertible a SqlValue => FieldDefinition -> a -> WhereCondition

(.<=) :: Convertible a SqlValue => FieldDefinition -> a -> WhereCondition

data SelectOptions

Constructors

SelectOptions

Fields

selectOptWhere :: [WhereCondition]

selectOptOrder :: [OrderByClause]

selectOptLimit :: First Int

selectOptOffset :: First Int

selectOptGroup :: [GroupByClause]

where_ :: WhereCondition -> SelectOptions

order :: ToOrderBy a => a -> SortDirection -> SelectOptions

limit :: Int -> SelectOptions

offset :: Int -> SelectOptions

groupBy :: ToGroupBy a => a -> SelectOptions

(<>) :: Semigroup a => a -> a -> a infixr 6

An associative operation.

data FieldUpdate

Constructors

FieldUpdate

Fields

fieldUpdateField :: FieldDefinition

fieldUpdateValue :: SqlValue

fieldUpdate :: Convertible a SqlValue => FieldDefinition -> a -> FieldUpdate

(.:=) :: Convertible a SqlValue => FieldDefinition -> a -> FieldUpdate

data OrderByClause

Constructors

OrderByClause String [SqlValue] SortDirection

data SortDirection

Constructors

Ascending

Descending

migrateSchema :: SchemaDefinition -> Orville ()

selectAll :: TableDefinition entity -> SelectOptions -> Orville [entity Record]

selectFirst :: TableDefinition entity -> SelectOptions -> Orville (Maybe (entity Record))

deleteRecord :: TableDefinition entity -> entity Record -> Orville ()

deleteWhere :: TableDefinition entity -> [WhereCondition] -> Orville Integer

findRecord :: TableDefinition entity -> Record -> Orville (Maybe (entity Record))

findRecords :: TableDefinition entity -> [Record] -> Orville (Map Record (entity Record))

findRecordsBy :: (Convertible SqlValue fieldValue, Ord fieldValue) => TableDefinition entity -> FieldDefinition -> SelectOptions -> Orville (Map fieldValue [entity Record])

insertRecord :: TableDefinition entity -> entity () -> Orville (entity Record)

insertRecordMany :: TableDefinition entity -> [entity ()] -> Orville ()

updateFields :: TableDefinition entity -> [FieldUpdate] -> [WhereCondition] -> Orville Integer

updateRecord :: TableDefinition entity -> Record -> entity key -> Orville (entity Record)

data MigrationError

Constructors

MigrationLockExcessiveRetryError String